+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- BREAKING CHANGE: Make many methods static
+
 ## [1.0.2] - 2024-12-18
 
 ### Fixed
@@ -13,5 +19,6 @@
 - Initial code taken over from gg\_json
 - Minor changes
 
+[Unreleased]: https://github.com/inlavigo/gg_direct_json/compare/1.0.2...HEAD
 [1.0.2]: https://github.com/inlavigo/gg_direct_json/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/inlavigo/gg_direct_json/tag/%tag